Ajuste en el tipo de numero de cuenta

+++ ms-cuentas/src/main/java/com/banco/digital/ms_cuentas/model/Account.java
@@ -17,7 +17,7 @@ public class Account {
     private String accountNumber;
 
     @Column(name = "persnum", nullable = false)
-    private Long personNumber;
+    private Integer personNumber;
 
     @ManyToOne
     @JoinColumn(name = "divisa")
