+++ ms-cuentas/src/main/java/com/banco/digital/ms_cuentas/service/AccountService.java
@@ -6,12 +6,12 @@ import com.banco.digital.ms_cuentas.model.CurrencyCode;
 
 import java.math.BigDecimal;
 
-public interface AccountService extends BaseService<Account, Long> {
+public interface AccountService extends BaseService<Account, Integer> {
     void save(Account account);
 
     boolean existsByAccountNumber(String accountNumber);
 
-    Account generateAccount(Long persNum, CurrencyCode currencyCode, AccountStatus accountStatus, BigDecimal salary);
+    Account generateAccount(Integer persNum, CurrencyCode currencyCode, AccountStatus accountStatus, BigDecimal salary);
 
     String generateAccountNumber();
 }
